requirements.txt
psycopg[binary]==3.1.8
azure-storage-blob==12.18.1
python-dotenv==1.0.0

